This is how I got SSH running in fargate
This includes tcpdump, although being able to poke around /proc was
enough to work out what was wrong with the egress proxies.

You will need to edit the container privileges in the task definition to
be able to do useful stuff like tcpdump

I wanted to get this into a side car ssh container but haven't had time
- also I'm not sure if you would be able to see the other containers to
well. So this is really a pattern for how to add SSH to an existing
container.